Rename files and folder, and change namespaces
